refactor(grow): extract LLM helper to llm_helper.py
Move _grow_llm_call(), _build_grow_error_feedback(),
_validate_and_insert_gaps(), and GapInsertionReport to a new
_LLMHelperMixin class in llm_helper.py. GrowStage now inherits
from _LLMHelperMixin, keeping all self._* references working.

Update test mock patch targets from grow.stage.* to
grow.llm_helper.* for with_structured_output, extract_tokens,
and _get_prompts_path.

Part of #851.
